deleted everything except basic login features
