--- lib/features/auth/errors/errors.dart
@@ -1,4 +1,0 @@
-export 'login.dart';
-
-// For the sake of simplicity share the same error models between
-// the auth services and usecases.
