--- lib/features/home/page/index.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-import 'layouts/mobile.dart';
-
-class HomePage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return HomePageLayoutMobile();
-  }
-}
