--- lib/features/home/page/widgets/icons.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomeIcon extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Icon(Icons.home);
-  }
-}
-
-class CourtIcon extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Icon(Icons.calendar_today);
-  }
-}
-
-class ProfileIcon extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Icon(Icons.person_outline);
-  }
-}
